Refact: Move some boilerplate code into an AbstractBrowser class
