+++ src/main/java/io/github/arthurc/autotest/lifecycle/Lifecycle.java
@@ -156,7 +156,6 @@ public abstract class Lifecycle {
 	/**
 	 * Calls an action within the lifecycle. The lifecycle publishes events before and after the action is called along
 	 * with the result of the action.
-	 * .
 	 *
 	 * @param action The action to call.
 	 * @param <T>    The type of the result.
